Move the PDF chart colors into the shared theme.
The Search Funnel and All Traffic widgets defined their line
colors as hardcoded hex values that duplicated tokens already in
`PDF_COLORS`. Each color now references its `PDF_COLORS` token in
`pdf-theme.ts` directly, so the hex values live in one place.

This also moves the donut palette. It adds the donut colors to
`PDF_COLORS`, moves `PIE_CHART_COLORS` into `pdf-theme.ts`, and
deletes `pie-chart-colors.ts`. All PDF color data now lives in the
theme.

`PDFLink` builds its styles with `createPDFStyles`, so it scales
with the page like the other PDF components.
